refactor: extract ClassicBrowserTabService from ClassicBrowserService
Extract tab management logic into a dedicated ClassicBrowserTabService that orchestrates
tab creation, switching, and closing operations. The main ClassicBrowserService now acts
as a pure facade, delegating tab operations to the new service while maintaining tab
group management logic that requires additional dependencies.

Key improvements:
- Created ClassicBrowserTabService with all tab management logic
- Made createTabWithState public to allow direct calls from event handlers
- Introduced handlePostTabCreation helper to consolidate tab group logic
- Removed code duplication by having a single source of truth for tab operations
- Maintained clean separation of concerns with each service having a single responsibility

The facade now purely coordinates between services without duplicating business logic.
