+++ services/browser/ClassicBrowserService.ts
@@ -11,6 +11,7 @@ import { EventEmitter } from 'events';
 import { ClassicBrowserViewManager } from './ClassicBrowserViewManager';
 import { ClassicBrowserStateService } from './ClassicBrowserStateService';
 import { ClassicBrowserNavigationService } from './ClassicBrowserNavigationService';
+import { ClassicBrowserTabService } from './ClassicBrowserTabService';
 
 // Default URL for new tabs
 const DEFAULT_NEW_TAB_URL = 'https://www.are.na';
@@ -30,6 +31,7 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
   private viewManager: ClassicBrowserViewManager;
   private stateService: ClassicBrowserStateService;
   private navigationService: ClassicBrowserNavigationService;
+  private tabService: ClassicBrowserTabService;
   private prefetchViews: Map<string, WebContentsView> = new Map();
   private cleanupInterval: NodeJS.Timeout | null = null;
   private snapshots: Map<string, string> = new Map();
@@ -64,6 +66,13 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
       viewManager: this.viewManager,
       stateService: this.stateService,
       eventEmitter: this.eventEmitter
+    });
+    
+    // Initialize the tab service
+    this.tabService = new ClassicBrowserTabService({
+      stateService: this.stateService,
+      viewManager: this.viewManager,
+      navigationService: this.navigationService
     });
     
     this.setupEventListeners();
@@ -306,7 +315,8 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
       this.logDebug(`windowId ${windowId}: ${details.disposition} detected, creating ${makeActive ? 'active' : 'background'} tab`);
       
       try {
-        this.createTabWithState(windowId, details.url, makeActive);
+        this.tabService.createTabWithState(windowId, details.url, makeActive);
+        this.handlePostTabCreation(windowId);
         this.logInfo(`windowId ${windowId}: Created ${makeActive ? 'active' : 'background'} tab for ${details.url}`);
       } catch (err) {
         this.logError(`windowId ${windowId}: Failed to create new tab:`, err);
@@ -332,7 +342,8 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
         this.logDebug(`windowId ${windowId}: Intercepted CMD+click via custom protocol for URL: ${targetUrl}`);
         
         try {
-          this.createTabWithState(windowId, targetUrl, false);
+          this.tabService.createTabWithState(windowId, targetUrl, false);
+          this.handlePostTabCreation(windowId);
           this.logInfo(`windowId ${windowId}: Created background tab for CMD+click to ${targetUrl}`);
         } catch (err) {
           this.logError(`windowId ${windowId}: Failed to create new tab for CMD+click:`, err);
@@ -574,76 +585,27 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
   }
 
   /**
-   * Creates a new tab in the browser window with control over whether it becomes active.
-   * @param windowId - The window to create the tab in
-   * @param url - Optional URL to load in the new tab
-   * @param makeActive - Whether to switch to the new tab (default: true for backward compatibility)
-   * @returns The ID of the newly created tab
-   */
-  private createTabWithState(windowId: string, url?: string, makeActive: boolean = true): string {
-    const browserState = this.stateService.states.get(windowId);
-    if (!browserState) {
-      throw new Error(`Browser window ${windowId} not found`);
-    }
-
-    // Create new tab
-    const tabId = uuidv4();
-    const newTab: TabState = {
-      id: tabId,
-      url: url || DEFAULT_NEW_TAB_URL,
-      title: 'New Tab',
-      faviconUrl: null,
-      isLoading: makeActive, // Only set loading if we're going to load it
-      canGoBack: false,
-      canGoForward: false,
-      error: null
-    };
-
-    // Add to tabs array
-    const state = this.stateService.states.get(windowId);
-    if (!state) throw new Error(`Browser window ${windowId} not found`);
-    state.tabs = [...state.tabs, newTab];
-    
-    // Only update active tab ID if requested
-    if (makeActive) {
-      if (state) state.activeTabId = tabId;
-      
-      // Load the new tab's URL into the WebContentsView to synchronize view with state
-      const view = this.viewManager.getView(windowId);
-      if (view && view.webContents && !view.webContents.isDestroyed()) {
-        const urlToLoad = newTab.url; // Use the URL from the tab we just created
-        // Use the navigation service's loadUrl method for consistent error handling
-        this.navigationService.loadUrl(windowId, urlToLoad).catch(err => {
-          this.logError(`[createTabWithState] Failed to load URL ${urlToLoad}:`, err);
-        });
-        this.logDebug(`[createTabWithState] Loading ${urlToLoad} in new active tab ${tabId}`);
-      }
-    } else {
-      this.logDebug(`[createTabWithState] Created background tab ${tabId} with URL ${newTab.url}`);
-    }
-    
-    // Send state update - if makeActive is false, don't change activeTabId
-    this.stateService.sendStateUpdate(windowId, makeActive ? newTab : undefined, makeActive ? tabId : undefined);
-    
-    // Check if we should create a tab group (2+ tabs)
-    this.checkAndCreateTabGroup(windowId);
-    
-    // Schedule tab group update
-    this.scheduleTabGroupUpdate(windowId);
-    
-    this.logDebug(`[createTabWithState] Created ${makeActive ? 'active' : 'background'} tab ${tabId} in window ${windowId}`);
-    return tabId;
-  }
-
-  /**
    * Creates a new tab in the browser window.
    * @param windowId - The window to create the tab in
    * @param url - Optional URL to load in the new tab
    * @returns The ID of the newly created tab
    */
   createTab(windowId: string, url?: string): string {
-    // Use the new helper method with makeActive = true for backward compatibility
-    return this.createTabWithState(windowId, url, true);
+    const tabId = this.tabService.createTab(windowId, url);
+    this.handlePostTabCreation(windowId);
+    return tabId;
+  }
+
+  /**
+   * Handles post-tab creation logic such as tab group management.
+   * This should be called after any tab creation operation.
+   * @param windowId - The window that had a tab created
+   */
+  private handlePostTabCreation(windowId: string): void {
+    // Check if we should create a tab group (2+ tabs)
+    this.checkAndCreateTabGroup(windowId);
+    // Schedule tab group update
+    this.scheduleTabGroupUpdate(windowId);
   }
 
   /**
@@ -652,70 +614,7 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
    * @param tabId - The ID of the tab to switch to
    */
   switchTab(windowId: string, tabId: string): void {
-    const browserState = this.stateService.states.get(windowId);
-    if (!browserState) {
-      throw new Error(`Browser window ${windowId} not found`);
-    }
-
-    const targetTab = browserState.tabs.find(t => t.id === tabId);
-    if (!targetTab) {
-      throw new Error(`Tab ${tabId} not found in window ${windowId}`);
-    }
-
-    // Save current tab's scroll position if we're switching away
-    const currentTab = browserState.tabs.find(t => t.id === browserState.activeTabId);
-    if (currentTab && currentTab.id !== tabId) {
-      const view = this.viewManager.getView(windowId);
-      if (view && view.webContents && !view.webContents.isDestroyed()) {
-        // Store scroll position for the current tab
-        view.webContents.executeJavaScript(`
-          ({ x: window.scrollX, y: window.scrollY })
-        `).then(scrollPos => {
-          (currentTab as TabState & { scrollPosition?: { x: number; y: number } }).scrollPosition = scrollPos;
-        }).catch(err => {
-          this.logDebug(`[switchTab] Failed to save scroll position: ${err}`);
-        });
-      }
-    }
-
-    // Update active tab
-    if (browserState) browserState.activeTabId = tabId;
-
-    // Load the new tab's URL in the WebContentsView
-    const view = this.viewManager.getView(windowId);
-    if (view && view.webContents && !view.webContents.isDestroyed()) {
-      if (targetTab.url && targetTab.url !== 'about:blank') {
-        // Use the navigation service's loadUrl method for consistent error handling
-        this.navigationService.loadUrl(windowId, targetTab.url).catch(err => {
-          this.logError(`[switchTab] Failed to load URL ${targetTab.url}:`, err);
-        });
-      } else {
-        // For new tabs or blank tabs, update the state immediately
-        const tabUpdate: Partial<TabState> = {
-          url: 'about:blank',
-          title: 'New Tab',
-          isLoading: false,
-          canGoBack: false,
-          canGoForward: false,
-          error: null
-        };
-        
-        // Apply the update to the target tab
-        const tabIndex = browserState.tabs.findIndex(t => t.id === tabId);
-        if (tabIndex !== -1) {
-          browserState.tabs = browserState.tabs.map((tab, i) => 
-            i === tabIndex ? { ...tab, ...tabUpdate } : tab
-          );
-        }
-      }
-    } else {
-      this.logWarn(`[switchTab] No valid view or webContents for window ${windowId}`);
-    }
-
-    // Send complete state update with updated tab info
-    this.stateService.sendStateUpdate(windowId, undefined, tabId);
-    
-    this.logDebug(`[switchTab] Switched to tab ${tabId} in window ${windowId}`);
+    this.tabService.switchTab(windowId, tabId);
   }
 
   /**
@@ -724,84 +623,9 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
    * @param tabId - The ID of the tab to close
    */
   closeTab(windowId: string, tabId: string): void {
-    const browserState = this.stateService.states.get(windowId);
-    if (!browserState) {
-      throw new Error(`Browser window ${windowId} not found`);
-    }
-
-    const tabIndex = browserState.tabs.findIndex(t => t.id === tabId);
-    if (tabIndex === -1) {
-      throw new Error(`Tab ${tabId} not found in window ${windowId}`);
-    }
-
-    // Don't close if it's the last tab - create a new one instead
-    if (browserState.tabs.length === 1) {
-      // Reset the last tab to a new tab state (create new array for immutability)
-      const newTabId = uuidv4();
-      const newTab: TabState = {
-        id: newTabId,
-        url: DEFAULT_NEW_TAB_URL,
-        title: 'New Tab',
-        faviconUrl: null,
-        isLoading: true, // Set to true since we're loading a real URL
-        canGoBack: false,
-        canGoForward: false,
-        error: null
-      };
-      const state = this.stateService.states.get(windowId);
-      if (state) state.tabs = [newTab];
-      if (state) state.activeTabId = newTabId;
-      
-      // Load the default URL into the WebContentsView
-      const view = this.viewManager.getView(windowId);
-      if (view && view.webContents && !view.webContents.isDestroyed()) {
-        // Use the navigation service's loadUrl method for consistent error handling
-        this.navigationService.loadUrl(windowId, DEFAULT_NEW_TAB_URL).catch(err => {
-          this.logError(`[closeTab] Failed to load default URL:`, err);
-        });
-      }
-      
-      // Send complete state update
-      this.stateService.sendStateUpdate(windowId, newTab, newTabId);
-      this.logDebug(`[closeTab] Replaced last tab with new tab ${newTabId} in window ${windowId}`);
-      return;
-    }
-
-    // Remove the tab
-    const state = this.stateService.states.get(windowId);
-    if (state) state.tabs = state.tabs.filter(t => t.id !== tabId);
-
-    // Determine the next active tab
-    const currentActiveTabId = state?.activeTabId;
-    let newActiveTabId = currentActiveTabId;
-    let newActiveTab: TabState | undefined;
-    
-    // If we're closing the active tab, determine which tab to activate
-    if (currentActiveTabId === tabId) {
-      const remainingTabs = state?.tabs || [];
-      // Try to switch to the tab that was next to the closed one
-      const newActiveIndex = Math.min(tabIndex, remainingTabs.length - 1);
-      newActiveTab = remainingTabs[newActiveIndex];
-      newActiveTabId = newActiveTab.id;
-      if (state) state.activeTabId = newActiveTabId;
-      
-      // Load the new active tab's URL into the WebContentsView
-      const view = this.viewManager.getView(windowId);
-      if (view && view.webContents && !view.webContents.isDestroyed() && newActiveTab && newActiveTab.url) {
-        // Use the navigation service's loadUrl method for consistent error handling
-        this.navigationService.loadUrl(windowId, newActiveTab.url).catch(err => {
-          this.logError(`[closeTab] Failed to load URL ${newActiveTab?.url}:`, err);
-        });
-      }
-    }
-    
-    // Send a single state update reflecting the complete new state
-    this.stateService.sendStateUpdate(windowId, newActiveTab, newActiveTabId);
-    
-    // Schedule tab group update
+    this.tabService.closeTab(windowId, tabId);
+    // Schedule tab group update after tab closure
     this.scheduleTabGroupUpdate(windowId);
-    
-    this.logDebug(`[closeTab] Closed tab ${tabId} in window ${windowId}, active tab is now ${newActiveTabId}`);
   }
 
   /**
@@ -1430,6 +1254,9 @@ export class ClassicBrowserService extends BaseService<ClassicBrowserServiceDeps
     
     // Clean up the navigation service
     await this.navigationService.cleanup();
+    
+    // Clean up the tab service
+    await this.tabService.cleanup();
     
     this.logInfo('[ClassicBrowserService] Service cleaned up');
   }
